requirements.txt
asgiref==3.8.1
Django==4.2.13
pillow==10.3.0
sqlparse==0.5.0
django-ckeditor==6.7.1
typing_extensions==4.12.2
